Change MemoryInfo.attributes type to uint8_t
PDI changed the type of `attributes` from uint8_t to size_t.
However the SMBIOS spec only reserved 1 byte for this field.
So we need to keep `attribute` in 1 byte otherwise the field after
`attribute` will be shifted.

Tested:
Before
    .MemorySizeInKB property u 1258291200

After
    .MemorySizeInKB property u 33554432

+++ include/dimm.hpp
@@ -168,7 +168,7 @@ struct MemoryInfo
     uint8_t serialNum;
     uint8_t assetTag;
     uint8_t partNum;
-    size_t attributes;
+    uint8_t attributes;
     uint32_t extendedSize;
     uint16_t confClockSpeed;
     uint16_t minimumVoltage;
